+++ Gemfile
@@ -94,3 +94,6 @@ gem 'payjp'
 
 gem 'dotenv-rails', groups: [:development, :test]
 
+gem 'gon'
+
+
