+++ app/javascript/application.js
@@ -2,3 +2,6 @@
 import "@hotwired/turbo-rails"
 import "controllers"
 import "item_price"
+import "pay"
+import "memo"
+
