--- app/javascript/packs/application.js
@@ -1,1 +1,0 @@
-import "./pay";
