+++ spec/factories/orders.rb
@@ -8,3 +8,4 @@ FactoryBot.define do
     phone_number { '09012345678' }
     token { 'tok_abcdefghijk00000000000000000' }
   end
+end
